ansible/roles/uptime-kuma/files/requirements.txt
python-socketio==5.6.0
requests==2.28.0
click==8.1.3
